rename grizzly.task to grizzly.tasks
renamed pip-generate-requirements.bash to pip-compile.bash
