+++ grizzly/steps/scenario/tasks.py
@@ -8,8 +8,8 @@ from behave import register_type, then  # pylint: disable=no-name-in-module
 from ..helpers import add_request_task
 from ...types import RequestDirection, RequestMethod
 from ...context import GrizzlyContext
-from ...task import PrintTask, WaitTask, TransformerTask, UntilRequestTask, DateTask
-from ...task.getter import getterof
+from ...tasks import PrintTask, WaitTask, TransformerTask, UntilRequestTask, DateTask
+from ...tasks.getter import getterof
 
 from grizzly_extras.transformer import TransformerContentType
