+++ requirements-dev.txt
@@ -153,7 +153,7 @@ jsonpath-ng==1.5.3
     # via grizzly-loadtester (setup.py)
 lazy-object-proxy==1.7.1
     # via astroid
-locust @ git+https://github.com/mgor/locust.git@bug/typing
+locust @ git+https://github.com/mgor/locust.git@fb4591cbe2074f9b8d37c7b588a8e20a25c714f8
     # via grizzly-loadtester (setup.py)
 lxml==4.8.0
     # via grizzly-loadtester (setup.py)
