+++ setup.py
@@ -39,8 +39,8 @@ setup(
         'influxdb>=5.3.1',
         'Jinja2>=3.0.3',
         'jsonpath-ng>=1.5.3',
-        #'locust==2.7.3',
-        'locust @ git+https://github.com/mgor/locust.git@bug/typing#egg=locust',
+        #'locust==2.8.3',
+        'locust @ git+https://github.com/mgor/locust.git@fb4591cbe2074f9b8d37c7b588a8e20a25c714f8#egg=locust',
         'lxml>=4.8.0',
         'mypy-extensions>=0.4.3',
         'opencensus-ext-azure>=1.1.1',
